refactor: model, package structure
