+++ src/main/kotlin/com/rootsid/wal/library/didcom/storage/SecretResolver.kt
@@ -1,4 +1,4 @@
-package com.rootsid.wal.library.didcom
+package com.rootsid.wal.library.didcom.storage
 
 import org.didcommx.didcomm.secret.Secret
 import org.didcommx.didcomm.secret.SecretResolverEditable
