+++ src/main/kotlin/com/rootsid/wal/library/wallet/model/Claim.kt
@@ -1,0 +1,32 @@
+package com.rootsid.wal.library.wallet.model
+
+import io.iohk.atala.prism.api.CredentialClaim
+import io.iohk.atala.prism.identity.PrismDid
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.json.Json
+
+/**
+ * Claim
+ *
+ * @property subjectDid
+ * @property content
+ * @constructor Create empty Claim
+ */
+@Serializable
+data class Claim(
+    val subjectDid: String,
+    val content: String
+)
+
+/**
+ * To credential claim
+ *
+ * Convert a Claim to PRISM CredentialClaim
+ */
+@OptIn(ExperimentalSerializationApi::class)
+fun Claim.toCredentialClaim() = CredentialClaim(
+    PrismDid.fromString(this.subjectDid),
+    Json.decodeFromString(this.content)
+)
