+++ src/main/kotlin/com/rootsid/wal/library/wallet/model/ImportedCredential.kt
@@ -1,0 +1,17 @@
+package com.rootsid.wal.library.wallet.model
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Imported credential
+ *
+ * @property alias
+ * @property verifiedCredential
+ * @constructor Create empty Imported credential
+ */
+@Serializable
+data class ImportedCredential(
+    val alias: String,
+    // Signed VC and proof (This is the real VC)
+    var verifiedCredential: VerifiedCredential,
+)
