+++ src/main/kotlin/com/rootsid/wal/library/wallet/model/Proof.kt
@@ -1,0 +1,18 @@
+package com.rootsid.wal.library.wallet.model
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Proof
+ *
+ * @property hash
+ * @property index
+ * @property siblings
+ * @constructor Create empty Proof
+ */
+@Serializable
+data class Proof(
+    var hash: String,
+    var index: Int,
+    var siblings: MutableList<String> = mutableListOf()
+)
